gpu and cpu build in linux (#12271)

+++ onnxruntime/core/framework/allocation_planner.cc
@@ -81,7 +81,6 @@ std::ostream& operator<<(std::ostream& out, AllocKind alloc_kind) {
 std::ostream& operator<<(std::ostream& out, std::pair<const SequentialExecutionPlan*, const SessionState*> planinfo) {
   const SequentialExecutionPlan& plan = *planinfo.first;
   const SessionState& session_state = *planinfo.second;
-  //auto& graph = session_state.GetGraphViewer();
 
   const auto& name_idx_map = session_state.GetOrtValueNameIdxMap();
   InlinedHashMap<int, std::string_view> index_to_name;
@@ -1737,7 +1736,7 @@ class PlannerImpl {
     plan_.node_release_list.resize(graph_viewer_.MaxNodeIndex() + 1);
     for (size_t i = 0; i < value_consumers.size(); ++i) {
       if (!value_consumers[i].empty()) {
-        plan_.release_actions.push_back(SequentialExecutionPlan::ReleaseAction{static_cast<int>(i), 0});  // TODO: narrowing conversion
+        plan_.release_actions.push_back(SequentialExecutionPlan::ReleaseAction{i, 0});  // TODO: narrowing conversion
         auto release_action_idx = plan_.release_actions.size() - 1;
         // check whether we can static determine where to release.
         // TODO: here we use a temporary simple solution is only static release when all the consumers are on the same stream
